requirements.txt
# Core Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.1.0
langchain-community>=0.3.0

# API Clients
openai>=1.12.0
elevenlabs>=1.0.0
replicate>=0.25.0

# Video Processing
moviepy>=1.0.3,<3.0.0  # Supports v1.x and v2.x (v2.2.1 is latest as of May 2025)
ffmpeg-python>=0.2.0
pillow>=10.2.0

# Web Framework
streamlit>=1.32.0
streamlit-player>=0.1.0

# State Management
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
tenacity>=8.2.0

# Data Handling
pydub>=0.25.1
numpy>=1.26.0
